# pup_patrol_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pup_patrol",
        "version": "1"
    },
    "paths": {
...
        "/api/comments/{postId}": {
            "get": {
                "tags": [
                    "comment"
                ],
                "summary": "특정 게시글의 댓글 반환",
                "description": "파라미터로 받은 postId와 comments 테이블의 postId열이 일치하는 정보들을 data라는 키를 가진 배열로 반환",
                "operationId": "d6810ca6c68ef6a0cd06244459be433d",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "post의 id",
                        "required": true,
                        "schema": {
                            "type": "json"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/comments": {
            "post": {
                "tags": [
                    "comment"
                ],
                "summary": "댓글 추가",
                "description": "comments 테이블에 데이터를 추가함",
                "operationId": "300619a0a4003920aafe960e45688102",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "commentPostId": {
                                        "description": "댓글이 달린 게시글 id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "commentContent": {
                                        "description": "댓글 내용",
                                        "type": "string",
                                        "example": "안녕하세요"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "answer이 성공적으로 추가됨"
                    }
                }
            }
        },
        "/api/comments/{comment}": {
            "get": {
                "tags": [
                    "comment"
                ],
                "summary": "파라미터로 넘어온 id를 가지는 특정 comment 데이터 조회",
                "description": "파라미터로 넘어온 id를 가지는 특정 comment 데이터를 data라는 키를 가진 객체로 반환",
                "operationId": "df604464ebd8be1727bb39c2711e2f03",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "comment의 id",
                        "required": true,
                        "schema": {
                            "type": "json"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "특정 댓글 데이터"
                    }
                }
            },
            "put": {
                "tags": [
                    "comment"
                ],
                "summary": "comment 데이터 수정",
                "description": "로그인 한 정보와 댓글 작성자가 같다면 comment 데이터를 수정함 ",
                "operationId": "11551b83dcd65f7f079adf59f8c0348e",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "comment의 아이디",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "commentPostId": {
                                        "description": "댓글이 달린 게시글 id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "commentContent": {
                                        "description": "댓글 내용",
                                        "type": "string",
                                        "example": "안녕하세요"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "수정이 된 경우"
                    },
                    "401": {
                        "description": "로그인된 정보와 댓글 작성자가 일치하지 않을 경우"
                    }
                }
            },
            "delete": {
                "tags": [
                    "comment"
                ],
                "summary": "특정 comment 데이터 삭제",
                "description": "로그인한 정보와 댓글 작성자가 같다면 comment를 삭제함",
                "operationId": "e2e2b84a31c09628af8301e2ab9b209c",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "comment의 id",
                        "required": true,
                        "schema": {
                            "type": "json"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "comment 삭제에 성공한 경우"
                    },
                    "401": {
                        "description": "로그인된 정보와 댓글 작성자가 일치하지 않을 경우"
                    }
                }
            }
        },
...
    },
...
}